@@ -3,33 +3,33 @@
   "version": "0.2.1",
   "private": true,
   "dependencies": {
-    "@monaco-editor/react": "^3.4.2",
+    "@monaco-editor/react": "^3.8.3",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
-    "connected-react-router": "^6.8.0",
     "corewar": "workspace:*",
     "history": "^4.10.1",
     "identicon.js": "^2.3.3",
     "jssha": "^3.1.0",
+    "monaco-editor": "^0.21.3",
     "prop-types": "^15.7.2",
     "pubsub-js": "^1.9.5",
     "random-words": "^1.1.1",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-redux": "^7.2.0",
-    "react-router-dom": "^5.2.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-redux": "^9.2.0",
+    "react-router-dom": "^6.30.1",
     "react-scripts": "5.0.1",
-    "redux": "^4.0.5",
+    "redux": "^5.0.1",
     "redux-saga": "^1.1.3",
     "tailwindcss": "^1.6.0",
     "typeface-anonymous-pro": "0.0.72",
     "typeface-inter": "^3.12.0"
   },
   "scripts": {
-    "dev": "react-scripts start",
-    "start": "react-scripts start",
-    "build": "react-scripts build",
+    "dev": "PORT=3000 react-scripts start",
+    "start": "PORT=3000 react-scripts start",
+    "build": "PUBLIC_URL=/app react-scripts build",
     "test": "react-scripts test --env=jsdom --passWithNoTests",
     "coverage": "react-scripts test --env=jsdom --coverage",
     "eject": "react-scripts eject",
