@@ -1,17 +1,25 @@
-import React from 'react'
 import PropTypes from 'prop-types'
+import React from 'react'
 import { Provider } from 'react-redux'
-import { Route } from 'react-router-dom'
-import { ConnectedRouter } from 'connected-react-router'
+import { BrowserRouter, Route, Routes } from 'react-router-dom'
 import App from './App'
 
-const Root = ({ store, history }) => (
-  <Provider store={store}>
-    <ConnectedRouter history={history}>
-      <Route component={App} />
-    </ConnectedRouter>
-  </Provider>
-)
+const Root = ({ store }) => {
+  return (
+    <Provider store={store}>
+      <BrowserRouter
+        future={{
+          v7_startTransition: true,
+          v7_relativeSplatPath: true
+        }}
+      >
+        <Routes>
+          <Route path="*" element={<App />} />
+        </Routes>
+      </BrowserRouter>
+    </Provider>
+  )
+}
 
 Root.propTypes = {
   store: PropTypes.object.isRequired
