@@ -1,5 +1,5 @@
 import React from 'react'
-import { useHistory, useParams } from 'react-router-dom'
+import { useNavigate, useParams } from 'react-router-dom'
 import FileBrowser from '../features/files/file-browser'
 import SimulatorSettings from '../features/simulator/simulator-settings'
 
@@ -15,12 +15,12 @@ const getContent = ({ id }) => {
 }
 
 const Modal = () => {
-  const history = useHistory()
+  const navigate = useNavigate()
   const content = getContent(useParams())
   return (
     <div
       className="fixed bottom-0 inset-x-0 z-20 px-4 pb-4 sm:inset-0 sm:flex sm:items-center sm:justify-center"
-      onClick={() => history.goBack()}
+      onClick={() => navigate(-1)}
     >
       <div className="fixed inset-0 transition-opacity">
         <div className="absolute inset-0 bg-gray-800 opacity-75"></div>
@@ -30,7 +30,7 @@ const Modal = () => {
         role="dialog"
         aria-modal="true"
         aria-labelledby="modal-headline"
-        onClick={e => e.stopPropagation()}
+        onClick={(e) => e.stopPropagation()}
       >
         {content}
       </div>
