@@ -1,12 +1,12 @@
 import React from 'react'
-import { useSelector, useDispatch } from 'react-redux'
-import { useHistory } from 'react-router-dom'
+import { useDispatch, useSelector } from 'react-redux'
+import { useNavigate } from 'react-router-dom'
 import { openFile } from './actions'
 
 const FileBrowser = () => {
-  const { warriorLibrary } = useSelector(state => state.file)
+  const { warriorLibrary } = useSelector((state) => state.file)
   const dispatch = useDispatch()
-  const history = useHistory()
+  const navigate = useNavigate()
 
   return (
     <div className="bg-gray-700 p-4">
@@ -15,13 +15,13 @@ const FileBrowser = () => {
           File Browser
         </h3>
         <ul>
-          {warriorLibrary.map(warrior => (
+          {warriorLibrary.map((warrior) => (
             <li
               key={warrior.name}
               className="flex items-center my-4 p-4 w-full text-sm rounded-lg border border-gray-600 cursor-pointer hover:bg-gray-600"
               onClick={() => {
                 dispatch(openFile(warrior.source))
-                history.goBack()
+                navigate(-1)
               }}
             >
               {warrior.name}
