@@ -369,41 +369,45 @@ const FeatureImageWrapper = styled.div`
   `}
 `
 
-const Home = () => (
-  <HomeGrid>
-    <SiteNav />
-    <HeroLogo />
-    <CallToAction>
-      <PrimaryButton href="/app">Play Now</PrimaryButton>
-      <FeatureButton href="/learn">Learn more</FeatureButton>
-    </CallToAction>
-    <Tagline id={`features`}>
-      <h2>
-        This is Corewar where players test their coding skills against each other, writing warriors
-        which battle for control of the core. Do you have what it takes to become the King of the
-        Hill?
-      </h2>
-    </Tagline>
-    <Simulator>
-      <FeatureDescription>
-        Our interactive Corewar Simulator provides the battle ground for your warriors. Play the
-        game, inspect the core and debug your code from your web browser!
-      </FeatureDescription>
-      <FeatureImageWrapper>
-        <img src={SimulatorImage} alt={`Animated core simulator example`} />
-      </FeatureImageWrapper>
-    </Simulator>
-    <Parser>
-      <FeatureImageWrapper>
-        <img src={ParserImage} alt={`Animated parser usage example`} />
-      </FeatureImageWrapper>
-      <FeatureDescription>
-        Our powerful Redcode editor makes it easy to write warriors and through real-time code
-        analysis improve the quality of your programs. The perfect canvas on which to craft your
-        masterpiece.
-      </FeatureDescription>
-    </Parser>
-    {/* <Features>
+const Home = () => {
+  const appUrl =
+    process.env.NODE_ENV === 'development' ? 'http://localhost:3000/editor' : '/app/editor'
+
+  return (
+    <HomeGrid>
+      <SiteNav />
+      <HeroLogo />
+      <CallToAction>
+        <PrimaryButton href={appUrl}>Play Now</PrimaryButton>
+        <FeatureButton href="/learn">Learn more</FeatureButton>
+      </CallToAction>
+      <Tagline id={`features`}>
+        <h2>
+          This is Corewar where players test their coding skills against each other, writing
+          warriors which battle for control of the core. Do you have what it takes to become the
+          King of the Hill?
+        </h2>
+      </Tagline>
+      <Simulator>
+        <FeatureDescription>
+          Our interactive Corewar Simulator provides the battle ground for your warriors. Play the
+          game, inspect the core and debug your code from your web browser!
+        </FeatureDescription>
+        <FeatureImageWrapper>
+          <img src={SimulatorImage} alt={`Animated core simulator example`} />
+        </FeatureImageWrapper>
+      </Simulator>
+      <Parser>
+        <FeatureImageWrapper>
+          <img src={ParserImage} alt={`Animated parser usage example`} />
+        </FeatureImageWrapper>
+        <FeatureDescription>
+          Our powerful Redcode editor makes it easy to write warriors and through real-time code
+          analysis improve the quality of your programs. The perfect canvas on which to craft your
+          masterpiece.
+        </FeatureDescription>
+      </Parser>
+      {/* <Features>
       <Feature>
         <Octicon name='rocket' />
         <h3>Play corewar</h3>
@@ -423,141 +427,143 @@ const Home = () => (
         <FeatureButton href='#roadmap'>View Roadmap</FeatureButton>
       </Feature>
     </Features> */}
-    {/* <Prospect id={`who-are-you`}>
+      {/* <Prospect id={`who-are-you`}>
       <h2>What best describes you?</h2>
       <Octicon name='person' />
     </Prospect> */}
-    <Features>
-      <Feature>
-        <Octicon name="mortar-board" />
-        <h3>Experienced player</h3>
-        <p>I’ve played corewar before and understand the instructions and concepts</p>
-        <PrimaryButton href="/app">Play Now</PrimaryButton>
-        <Guidance>
-          <Octicon name="hubot" />
-          <SpeechArrow />
-          <SpeechBubble>
-            You should head over to the app and follow the interactive guide to explore the features
-          </SpeechBubble>
-        </Guidance>
-      </Feature>
-      <Feature>
-        <Octicon name="law" />
-        <h3>New to the game</h3>
-        <p>I’ve done some coding before but never played corewar</p>
-        <FeatureButton href="/learn">View Tutorial</FeatureButton>
-        <Guidance>
-          <Octicon name="hubot" />
-          <SpeechArrow />
-          <SpeechBubble>
-            You should run through the tutorial to understand the basics of corewar first, then head
-            over the app to test what you've learned.
-          </SpeechBubble>
-        </Guidance>
-      </Feature>
-    </Features>
-    <Prospect id={`roadmap`}>
-      <h2>The roadmap</h2>
-      <Octicon name="milestone" />
-    </Prospect>
-    <Roadmap>
-      <RoadmapItem>
-        <RoadmapText left>November 2017</RoadmapText>
-        <IconWrapper colour={colour.blue}>
-          <Octicon name="light-bulb" />
-        </IconWrapper>
-        <RoadmapText right>
-          The idea to develop corewar as a modern web app is concieved
-        </RoadmapText>
-        <Timeline />
-      </RoadmapItem>
+      <Features>
+        <Feature>
+          <Octicon name="mortar-board" />
+          <h3>Experienced player</h3>
+          <p>I’ve played corewar before and understand the instructions and concepts</p>
+          <PrimaryButton href={appUrl}>Play Now</PrimaryButton>
+          <Guidance>
+            <Octicon name="hubot" />
+            <SpeechArrow />
+            <SpeechBubble>
+              You should head over to the app and follow the interactive guide to explore the
+              features
+            </SpeechBubble>
+          </Guidance>
+        </Feature>
+        <Feature>
+          <Octicon name="law" />
+          <h3>New to the game</h3>
+          <p>I’ve done some coding before but never played corewar</p>
+          <FeatureButton href="/learn">View Tutorial</FeatureButton>
+          <Guidance>
+            <Octicon name="hubot" />
+            <SpeechArrow />
+            <SpeechBubble>
+              You should run through the tutorial to understand the basics of corewar first, then
+              head over the app to test what you've learned.
+            </SpeechBubble>
+          </Guidance>
+        </Feature>
+      </Features>
+      <Prospect id={`roadmap`}>
+        <h2>The roadmap</h2>
+        <Octicon name="milestone" />
+      </Prospect>
+      <Roadmap>
+        <RoadmapItem>
+          <RoadmapText left>November 2017</RoadmapText>
+          <IconWrapper colour={colour.blue}>
+            <Octicon name="light-bulb" />
+          </IconWrapper>
+          <RoadmapText right>
+            The idea to develop corewar as a modern web app is concieved
+          </RoadmapText>
+          <Timeline />
+        </RoadmapItem>
 
-      <RoadmapItem>
-        <RoadmapText left>March 2018</RoadmapText>
-        <IconWrapper colour={colour.blue}>
-          <Octicon name="calendar" />
-        </IconWrapper>
-        <RoadmapText right>
-          Public beta, allowing players to experience the initial feature set
-        </RoadmapText>
-        <Timeline />
-      </RoadmapItem>
+        <RoadmapItem>
+          <RoadmapText left>March 2018</RoadmapText>
+          <IconWrapper colour={colour.blue}>
+            <Octicon name="calendar" />
+          </IconWrapper>
+          <RoadmapText right>
+            Public beta, allowing players to experience the initial feature set
+          </RoadmapText>
+          <Timeline />
+        </RoadmapItem>
 
-      <RoadmapItem>
-        <RoadmapText left>Today</RoadmapText>
-        <IconWrapper colour={colour.blue}>
-          <Octicon name="megaphone" />
-        </IconWrapper>
-        <RoadmapText right>
-          We are now actively listening for user feedback whilst we build the next set of features
-          but here’s what we had in mind
-        </RoadmapText>
-        <Timeline />
-      </RoadmapItem>
+        <RoadmapItem>
+          <RoadmapText left>Today</RoadmapText>
+          <IconWrapper colour={colour.blue}>
+            <Octicon name="megaphone" />
+          </IconWrapper>
+          <RoadmapText right>
+            We are now actively listening for user feedback whilst we build the next set of features
+            but here’s what we had in mind
+          </RoadmapText>
+          <Timeline />
+        </RoadmapItem>
 
-      <TimelineDivider>
-        <h3>Future development</h3>
-      </TimelineDivider>
+        <TimelineDivider>
+          <h3>Future development</h3>
+        </TimelineDivider>
 
-      <RoadmapItem>
-        <RoadmapText left>Log in &amp; Warrior Management</RoadmapText>
-        <IconWrapper colour={colour.blue}>
-          <Octicon name="git-commit" />
-        </IconWrapper>
-        <RoadmapText right>
-          Create an account and build up your warriors, including file versioning and statistics
-          over time
-        </RoadmapText>
-        <Timeline />
-      </RoadmapItem>
+        <RoadmapItem>
+          <RoadmapText left>Log in &amp; Warrior Management</RoadmapText>
+          <IconWrapper colour={colour.blue}>
+            <Octicon name="git-commit" />
+          </IconWrapper>
+          <RoadmapText right>
+            Create an account and build up your warriors, including file versioning and statistics
+            over time
+          </RoadmapText>
+          <Timeline />
+        </RoadmapItem>
 
-      <RoadmapItem>
-        <RoadmapText left>Live online hills</RoadmapText>
-        <IconWrapper colour={colour.blue}>
-          <Octicon name="git-commit" />
-        </IconWrapper>
-        <RoadmapText right>
-          We will host and run hills for all skill levels. Which will allow you to submit you
-          warriors online and see the results in realtime
-        </RoadmapText>
-        <Timeline />
-      </RoadmapItem>
+        <RoadmapItem>
+          <RoadmapText left>Live online hills</RoadmapText>
+          <IconWrapper colour={colour.blue}>
+            <Octicon name="git-commit" />
+          </IconWrapper>
+          <RoadmapText right>
+            We will host and run hills for all skill levels. Which will allow you to submit you
+            warriors online and see the results in realtime
+          </RoadmapText>
+          <Timeline />
+        </RoadmapItem>
 
-      <RoadmapItem>
-        <RoadmapText left>Private hills</RoadmapText>
-        <IconWrapper colour={colour.blue}>
-          <Octicon name="git-commit" />
-        </IconWrapper>
-        <RoadmapText right>
-          Want to play with just your friends? We will add invite only hills so you can chose the
-          rules and the players
-        </RoadmapText>
-        <Timeline />
-      </RoadmapItem>
+        <RoadmapItem>
+          <RoadmapText left>Private hills</RoadmapText>
+          <IconWrapper colour={colour.blue}>
+            <Octicon name="git-commit" />
+          </IconWrapper>
+          <RoadmapText right>
+            Want to play with just your friends? We will add invite only hills so you can chose the
+            rules and the players
+          </RoadmapText>
+          <Timeline />
+        </RoadmapItem>
 
-      <RoadmapItem>
-        <RoadmapText left>Corewar challenges</RoadmapText>
-        <IconWrapper colour={colour.blue}>
-          <Octicon name="git-commit" />
-        </IconWrapper>
-        <RoadmapText right>
-          Feel like you’ve done all corewar can offer? We will develop specific challenges and
-          situations to challenge the most experienced players
-        </RoadmapText>
-        <Timeline />
-      </RoadmapItem>
-    </Roadmap>
-    <Prospect>
-      <h2>Check out the app</h2>
-      <Octicon name="beaker" />
-      <PrimaryButton href="/app">Play Now</PrimaryButton>
-    </Prospect>
-    <Footer>
-      &copy; 2018 <a href="http://www.corewar.io">www.corewar.io</a> - crafted with love by{' '}
-      <a href="https://github.com/gareththegeek">@gareththegeek</a> &amp;{' '}
-      <a href="https://github.com/dougajmcdonald/">@dougajmcdonald</a>
-    </Footer>
-  </HomeGrid>
-)
+        <RoadmapItem>
+          <RoadmapText left>Corewar challenges</RoadmapText>
+          <IconWrapper colour={colour.blue}>
+            <Octicon name="git-commit" />
+          </IconWrapper>
+          <RoadmapText right>
+            Feel like you’ve done all corewar can offer? We will develop specific challenges and
+            situations to challenge the most experienced players
+          </RoadmapText>
+          <Timeline />
+        </RoadmapItem>
+      </Roadmap>
+      <Prospect>
+        <h2>Check out the app</h2>
+        <Octicon name="beaker" />
+        <PrimaryButton href={appUrl}>Play Now</PrimaryButton>
+      </Prospect>
+      <Footer>
+        &copy; 2018 <a href="http://www.corewar.io">www.corewar.io</a> - crafted with love by{' '}
+        <a href="https://github.com/gareththegeek">@gareththegeek</a> &amp;{' '}
+        <a href="https://github.com/dougajmcdonald/">@dougajmcdonald</a>
+      </Footer>
+    </HomeGrid>
+  )
+}
 
 export default Home
